08NeuroHub/requirements.txt
# Python渚濊禆鍖呭垪琛?
# 妯＄粍鍏€绘帶妯″潡鎵€闇€鐨凱ython鍖?

# Web妗嗘灦
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 娑堟伅闃熷垪
pyzmq==25.1.1

# 鏁版嵁搴?
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# 鏁版嵁澶勭悊
pandas==2.1.3
numpy==1.25.2
pyyaml==6.0.1

# 寮傛鏀寔
aioredis==2.0.1
aiofiles==23.2.1

# 鐩戞帶鍜屾棩蹇?
prometheus-client==0.19.0
structlog==23.2.0

# 瀹夊叏
cryptography==41.0.7
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# HTTP瀹㈡埛绔?
httpx==0.25.2
aiohttp==3.9.1

# 閰嶇疆绠＄悊
pydantic==2.5.0
pydantic-settings==2.1.0

# 宸ュ叿搴?
click==8.1.7
rich==13.7.0
typer==0.9.0

# 娴嬭瘯 (寮€鍙戠幆澧?
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# 寮€鍙戝伐鍏?
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# 鏃堕棿澶勭悊
python-dateutil==2.8.2
pytz==2023.3

# 搴忓垪鍖?
orjson==3.9.9
msgpack==1.0.7
